refactor: move masking evaluator to a separate file
